lxd/storage/block: Move utility functions from connector package
